# Optimize preprocessing pipeline