## What

Optimize the preprocessing pipeline by using TF ops instead of python + adding cache.

The full training step on Train.ipynb now takes about 2 hours on my machine (laptop, CPU training), vs 7-8 hours IIRC.


## Changes

- Convert TFTransformer to tf.data.Dataset pipeline ops.
- Enable Dataset file-based caching.

## Benchmark

Benchmark on a 3-epoch test run:
- Reference: 52min 23s, ~1000s/epoch
- Optimized: 23min 13s, ~550s/1st epoch, then ~400s/epoch

Legend:
- Orange: Reference
- Blue: Reference with cache enabled
- Red: Optimized

![tensorboard-benchmark](https://user-images.githubusercontent.com/30210525/164107552-8044d450-4b7d-468a-a69e-cf3c17391a04.png)

### Tensorboard profile for Reference

<img src="https://user-images.githubusercontent.com/30210525/164107599-b71dd661-32fa-4d9e-930f-6950d00ef7db.png" width="400">

Cost: Dataset (31.3%) + PyFunc (26%) = 57%

### Tensorboard profile for Optimized

<img src="https://user-images.githubusercontent.com/30210525/164107635-70bc29bf-1fe7-401a-81eb-5a97d709cd3b.png" width="400">

Cost: Dataset (19.7%) + PyFunc (7.7%) = 25.5%