@streino 2022-04-19
The PR works but I get warnings:

    W tensorflow/core/kernels/data/cache_dataset_ops.cc:233] The calling iterator did not fully read the dataset being cached. In order to avoid unexpected truncation of the dataset, the partially cached contents of the dataset  will be discarded. This can happen if you have an input pipeline similar to `dataset.cache().take(k).repeat()`. You should use `dataset.take(k).cache().repeat()` instead.

This one shows up a couple times in the first epochs, then stops (could be a limit to the number of identical messages in the log).
I don't see anything wrong in our code that could cause this, and there's an open bug on TF that seems relevant: https://github.com/tensorflow/tensorflow/issues/50952

    W tensorflow/python/util/util.cc:348] Sets are not currently considered sequences, but this may change in the future, so consider avoiding using them.

That one's a TF bug: https://github.com/tensorflow/tensorflow/issues/48561

I think the massive perf improvements are worth a couple warnings, but let me know if you disagree.