@8huit 2022-05-03
Cheers! These are great improvements!
It runs for 3h on my macbook and on colab (tpu) vs 8h before!